Uni_UZH/WS23:24/ltwa_bullinger/bullinger-project-new/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg[binary]==3.1.10
